requirements.txt
bcrypt==3.2.0
blinker==1.8.2
boto3==1.21.7
botocore==1.24.46
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
Flask==3.0.3
idna==3.8
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
packaging==24.0
pluggy==1.4.0
pre-commit-hooks==4.5.0
pycparser==2.22
pytest==8.1.1
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.26.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.5.2
schedule==1.2.1
six==1.16.0
timedelta==2020.12.3
urllib3==1.26.20
Werkzeug==3.0.3
